+++ apps/api/src/admin/modules/admins/providers.ts
@@ -1,5 +1,6 @@
 import { type Provider } from '@nestjs/common';
 
 import { AdminAdminsService } from './admin-admins.service';
+import { StripeWebhookService } from '../../../consumer/modules/payment-methods/stripe-webhook.service';
 
-export const providers = [AdminAdminsService] satisfies Provider[];
+export const providers = [StripeWebhookService, AdminAdminsService] satisfies Provider[];