+++ apps/api/src/consumer/modules/payments/consumer-payments.service.ts
@@ -268,14 +268,10 @@ export class ConsumerPaymentsService {
         amount: true,
       },
     });
-    console.log(`rows`, rows);
 
     const result: Record<$Enums.CurrencyCode, number> = {} as any;
 
     for (const row of rows) {
-      console.log(`\n************************************`);
-      console.log(`row`, row);
-      console.log(`************************************\n`);
       result[row.currencyCode] = Number(row._sum.amount ?? 0);
     }